Move file mixin spec
